Fixes nette/http#16 - SessionSection notice when unlimited session lifetime is set
  * includes test and fix for Notice: The expiration time is greater than the session expiration

+++ src/Http/SessionSection.php
@@ -189,7 +189,7 @@ class SessionSection extends Nette\Object implements \IteratorAggregate, \ArrayA
 			$whenBrowserIsClosed = TRUE;
 		} else {
 			$time = Nette\Utils\DateTime::from($time)->format('U');
-			$max = ini_get('session.gc_maxlifetime');
+			$max = (int) ini_get('session.gc_maxlifetime');
 			if ($max !== 0 && ($time - time() > $max + 3)) { // 0 - unlimited in memcache handler, 3 - bulgarian constant
 				trigger_error("The expiration time is greater than the session expiration $max seconds", E_USER_NOTICE);
 			}
